:truck: Moved content to separate __parts location
