+++ aksel.nav.no/website/components/website-modules/examples/__parts/Dekorator.tsx
@@ -8,28 +8,6 @@ function Footer() {
   return <div id="decorator-footer" />;
 }
 
-function Content({ children }: { children: React.ReactNode }) {
-  return (
-    <div className="decorator-content">
-      <style>
-        {`.decorator-content {
-          width: 100%;
-          padding-block: 5rem;
-          padding-inline: 2rem;
-          height: 100%;
-          background: repeating-linear-gradient(
-            45deg,
-            #eee,
-            #eee 10px,
-            #fff 10px,
-            #fff 20px
-          );`}
-      </style>
-      {children}
-    </div>
-  );
-}
-
 const MILJO_URL = "https://www.nav.no/dekoratoren";
 
 function Env() {
@@ -65,4 +43,4 @@ function useDekorator() {
   }, []);
 }
 
-export { Header, Content, Footer, Env, useDekorator };
+export { Header, Footer, Env, useDekorator };
