+++ aksel.nav.no/website/pages/eksempler/primitive-page/below-fold.tsx
@@ -1,7 +1,7 @@
 import { Heading, Page } from "@navikt/ds-react";
+import { Content } from "@/web/examples/__parts/PageDemoContent";
 import { withDsExample } from "@/web/examples/withDsExample";
 import {
-  Content,
   Env,
   Footer,
   Header,
